fix : 버튼, 인풋 폰트 미적용 및 프롬프트 삭제 수정
